requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==8.3.3
boto3==1.35.0
Pillow==10.3.0
python-jose[cryptography]
passlib[bcrypt]
